fix(routing): settings page routing
Before this commit the Settings Page was not considered as a route,
instead all the content of the page was replacing the content below,
since the `_notTransitionPageRoute()` method was used to "push" the
route.

After this commit to push the route the `_cupertinoPageRoute()` method
is being used. With that edit the Settings Page is treated like a normal
route, and so it's possible to swipe back and navigate correctly to the
route.

Closes: #218